Node bookmarks
This is a concept made in order to optimize paths with lots of nodes.
Every 500 pixels, there is a bookmark which points to the last node before that point.

For example:

0                       500                      1000                     1500
v                        v                        v                        v
|----|-------------|---|-|-|---|----------------------------|-----------|---|
0    1             2   3 4 5   6                            7           8   9

Node bookmarks:
0 -> 0
1 -> 4
2 -> 6
3 -> 8

This optimizes numerous calls to `Path:getNodeID()`, as there is no need to iterate through tens of nodes each time.

---

To get a bookmark number, we just need to floor divide a distance by 500.
Then, in the node ID function, we start from a node from that bookmark, instead of 0.