Collectible generators
Added collectible generators. They allow for better customization of 
drops, such as powerups, gems and coins.

One places the generators in a folder 
`<game>/config/collectible_generators/` and a listing file in 
`<game>/config/collectible_generators.json`.

The listing file looks like this:
```
[
  "vanilla_powerup",
  "vanilla_powerup_colorbomb"
]
```
It's simply a listing of the generators inside the 
`<game>/config/collectible_generators/` folder. File names must match 
with the entries inside this listing, without the `.json` format name.

An example powerup generator inside looks like this:
```
[
  [
    {"type":"powerup","name":"wild","weight":1000},
    {"type":"powerup","name":"fireball","weight":500},
    {"type":"powerup","name":"lightning","weight":500},
    {"type":"powerup","name":"stop","weight":1000},
    {"type":"powerup","name":"slow","weight":1000},
    {"type":"powerup","name":"reverse","weight":1000},
    {"type":"powerup","name":"speedshot","weight":1000},
    
{"type":"collectible_generator","name":"vanilla_powerup_colorbomb","weight":500}
  ],
  [
    {"type":"collectible_generator","name":"vanilla_gem"}
  ]
]
```

This is a list of pools. The generator goes through pools until it can 
find one, which can spawn a powerup. Each entry can be of two types: 
`powerup` and `collectible_generator`.
The further fields are as follows:
- `name` - A powerup name, or when another generator is specified, its 
name.
- `weight` - Optional, defaults to 1. When higher, the item is more 
likely to be picked.
- `conditions` - Optional, is a list. When specified, the item can be 
picked only when all conditions inside are met.
The only condition currently supported is
`{"type":"color_present","color":<color>}`, which checks whether the 
<color> is present.

Currently, all levels use the same powerup generator, which name is 
"vanilla_powerup". This will be changed in the future.