Added highscore data for UI string parsing
For parseable strings, a new compound type called `highscoreData` was 
added. It has two additional fields to get a specific data chunk:
`name` (string) - can be `"name"`, `"score"` or `"level"`. It's a 
particular piece of a single highscore entry. New possible data can be 
added in the future.
`place` (number) - Specifies from which place we are looking for. Note 
that you can't search for far places, because the data for them is not 
contained inside the Highscore data.

Example:
```
[
    "player: ",
    {
        "type": "highscoreData",
        "name": "name",
        "place": 1
    }
]
```
This will return a string `player: X` where `X` is the name of a 
currently leading player in the leaderboard.