More types to parsers and dynamic widget variables
String compound types:
- New `scoreFormat`: takes a `value` number and converts it into string 
so there are commas separating groups of digits.
- Fix `highscoreData`: Now the `place` parameter is actually parsed, so 
it can be derived from another provider.

Number compound types:
- Now when a string is provided where there should be a number, the 
string is automatically converted.
- For string providers, added `fromString` type which takes a `value` 
parameter, which can be a string provider or a string itself. Useful 
when string providers provide a string representation of a number, such 
as in `highscoreData` type.

Added a new field to widget data: you can now specify a `variables` 
field, which is a table of variables to append when drawing such Widget.
This works even on widgets which do not have a visual representation.