Hook widget actions to callbacks
Each JSON widget object now can have an additional field, `callbacks`.
It is an objects with any number of fields - each field name is a widget action, and the value is a name of the corresponding function in the UI script.

Example:
hud.json:
```
...
"Button_Pause": {
  ...
  "callbacks": {
    "hideEnd": "fun1"
  }
}
...
```

script.lua:
```
...
function fun1(f)
  -- do something here
end
...
```