Add more parameters to getDirListing()
`path` - a path to start with
`filter` can be "all", "file" or "dir". Filters results. Note: when "dir", directories are returned WITHOUT trailing / character, but in "all" mode, they still retain this character!
`recursive` - When set, this searches all subfolders.
`pathRec` is an internal use parameter and should not be used.