Add game documentation stub
For now, it contains information for four configuration files:
- gameplay
- highscores
- powerups
- spheres
A lot more is to be covered, after all it's just a stub.

lmao.json is going to be a file which will contain the raw documentation data.
This is because I plan to make a script which will generate html files automatically, based on that data.