! Bezier interpolation for level speeds
In levels/*.json files, in a speed node definition, a new optional field, `transition`, has been added.
It contains an array with a `type` field, which can be set to either `"linear"` or `"bezier"`.
If set to `"bezier"`, two more fields are required, `point1` and `point2`. These fields define the control points for Bezier interpolation between speeds.

Interpolation between speeds is set in the first node, so it applies between that and the next node.

Example:
```
{
    "distance": 0,
    "speed": 800,
    "transition": {
        "type": "bezier",
        "point1": 0.9,
        "point2": 0.9
    }
}
```