Add Expressions stub
Added Expressions, which can be created from a string which is just... well, an expression.
For example: `2+2` evaluates to `4` and `-(8/0.5)+3` evaluates to `-13`.
For now, expressions allow only numbers and four basic operations. No spaces are allowed.
You can test the expressions yourself by entering a command `expr <expression>` in the console.
Later on, with parameters, expressions will be a key part in doing advanced maths within the configuration files.