! Add Logger
Logger is an utility which aggregates most of the debug print statements
and saves them automatically to a text file.
The file is called `log.txt` and is overwritten each time the engine is run.
By default, it is being saved during a successful exit as well as every 10 seconds.
Quite a few other source files have been updated to support this feature.