! Zuma shooter
This commit adds support for a new shooter type, used in sphere matching
games like Zuma.
The circular shooter, as it is what it will be referred to, is for now
hardcoded to be active and the shooter is placed at the center
of the screen.
This commit also changes the way cut speed shot beams are rendered.

No changes to game files have been made yet.