! New data for Shooters
New appearance and sound data has been added to the Shooters.
Although all of the fields mentioned are mandatory by design,
the config class will prepend previously used values so the engine will
work on old data without any problems.
This is a good example of the compatibility layer in action, too.

Changes in shooter data:
- New fields were added, so the appearance of the shooter is much less hardocded.
  - `spriteOffset` (`{"x":0,"y":0}`) - The offset of the main shooter sprite.
  - `spriteAnchor` (`{"x":0.5,"y":0}`) - The anchor point of the main shooter sprite.
  - `shadowSpriteOffset` (`{"x":8,"y":8}`) - The offset of the shooter shadow sprite.
  - `shadowSpriteAnchor` (`{"x":0.5,"y":0}`) - The anchor point of the shooter shadow sprite.
  - `nextBallOffset` (`{"x":0,"y":21}`) - The offset of the next ball indicator sprite.
  - `nextBallAnchor` (`{"x":0.5,"y":0}`) - The anchor point of the next ball indicator sprite.
  - `sounds` - A table with two fields:
    - `sphereSwap` (`"sound_events/shooter_swap.json"`) - The sound event played when two spheres are swapped with each other.
    - `sphereFill` (`"sound_events/shooter_fill.json"`) - The sound event played when the shooter gets new spheres.
All of these fields are required; in order to retain the current behavior, use values in brackets.