! Add Ghost Spheres
Ghost spheres have been added in order to increase accuracy of Zuma's
sphere physics recreation and should not be used in other cases.
In other words, ghost spheres are spheres which appear to be destroyed,
but in reality they still exist on the board for certain time.
They have no hitbox, too.

Additionally, you can now adjust the knockback so it's constant for a
certain time.

Backported from ZR branch.

Changes in gameplay.json:
- Added two new fields in the `sphereBehaviour` section:
  - `knockbackTime` - Optional. If set, this will be the time during which the spheres' knockback speed will be "locked" and will not decrease over time.
  - `knockbackStopAfterTime` - Optional. If set to `true`, after the `knockbackTime` elapses, the spheres will stop abruptly rather than slowly deccelerating.

Changes in sphere effect data:
- Added a new field, `ghostTime`.
  - If set, the spheres will turn into ghosts for a set amount of time when destroyed using this effect.