Move documentation files out of the repository
The generated files are now stored in the doc/game/out folder.
The generate script does not reflect that change, as the HTML portion of
the generator is pending an overhaul.
This change has been made in order to not clutter the repository with
generated files.

Generated documentation will still be available in releases, and will
remain to be able to be generated via the Python script.