!!! New Resource Manager
The Resource Manager has got a big overhaul, which intends to simplify game file management by handling resources in a more organized way.
Right now, basic features are implemented, and there's a lot more to come!

There is a new way the resources are recognized. Instead of having to be in a specific folder, they now need to have an appropriate extension or, if it's a JSON file, a JSON schema assigned to it.
If something tries to get a resource that is not loaded, it is loaded first before erroring out. This also means that an initial load list is no longer necessary, so it has been removed.
A new feature, called load counters, is a way to incorporate more progress bars. Right now, it handles the game's progress bar, but in the future, it can be extended to support progress bars for maps and other sets of resources!
The resources contained in the root game folder, as well as in `config` and `maps` subfolders, are not loaded.

Known issues and caveats:
- Because both music and sound files have the same extension, `.ogg`, right now any files in the `music` subfolder will automatically be considered as a music file. Otherwise, it will be a sound file.
- Color Palettes have stopped working, so Wild powerups and gems will crash the game.

Changes in game data:
- The file `config/loadlist.json` has been removed.