! Add music track config
Music tracks are a successor to the `config/music.json` file, which has been now removed.
Essentially it's the same thing, just split across multiple files.
Right now, accessing the music files stays as it is - without specifying an extension or the `music_tracks/` folder. This behavior will be changed soon, though.

- A new resource type has been implemented: music tracks.
  - The data is an object with just one field:
    - `audio` - The audio file to be used for this music track.

- The file `config/music.json` has been removed.