! Remove 2D frame counts from Sprites
Sprites no longer support two-dimensional frame counts, as they were unnecessarily complicating the Sprite system and also were never used.
All sprite numbers are now of an `integer` type instead of `Vector2`.
Note that Sprite files still require the frame counts to be defined as a vector, as the frames can still be placed vertically on the image itself. However, to access the frames in sprites defined in such way, the frames will be now numbered left-to-right and then top-to-bottom, instead of using a two-dimensional index.

For example, a Sprite which has a state consisting of 4x3 frames, 12 frames total, would have the following indexes:
Vec2(1, 1) -> 1
Vec2(2, 1) -> 2
Vec2(3, 1) -> 3
Vec2(4, 1) -> 4
Vec2(1, 2) -> 5
Vec2(2, 2) -> 6
Vec2(3, 2) -> 7
Vec2(4, 2) -> 8
Vec2(1, 3) -> 9
Vec2(2, 3) -> 10
Vec2(3, 3) -> 11
Vec2(4, 3) -> 12

This change is also reflected in the Sprite Editor, which now also highlights the frames when hovered on the frame list.