Fix resource overwrite glitch
Whenever a resource was queued and then indirectly loaded, its entry would not be properly removed from the queue, due to a typo.
This meant that the queued resource would be loaded again, overwriting its previous instance.
The primary instance would still exist and be referenced in the place from which it was loaded first though, which had some unexpected consequences.
In this case, the configs of Path Entities had sound events loaded in such way and since they would be removed from the Resource Manager the moment their duplicates are loaded, they would not figure in the resource list, and thus not update.
Updating sound events is needed to adapt to the game volume. Hence, Path Entities' sounds did not follow the game volume.