! Config Class generation part 1
The DocLang generator now can generate simple Config Classes. For now, it will not overwrite anything, and it is in an early testing state. It has been tested and checked only with Path Entities and Score Events, and it is producing fully accurate results.

In order to test it (prints the result to the console), replace `main()` body with:
```
	path = "data/score_event.docl"
	docl_print_docld(path)
	docl_print_lua(path, "ScoreEventConfig")
	input()
```