! Config Class generation part 4
All of the Config Classes, excluding UI2, have been generated with the documentation generator and work just fine, with just one change (the `path` field in Sprite renamed to `image` - this will be changed soon in the actual sprite definitions as well).

Changes/additions:
- Vector2 is now `require`'d where it's needed. (default Vec2 parameter values)
- Default value for Objects: `{}` which means that the Config Class instance will still have this object (existing but empty) even if it's not present in the respective `.json` file.
- The generated class files no longer have an extra empty line at the end of file.
- UI2 files now have proper names, albeit hardcoded.

Currently, the output is still redirected to the `doc/game/out_lua` directory.