! Sunset UI2
UI2 was an attempt to make a better UI system in replacement of an old UI1 (also called just UI) which is still doing its heavy duty for now and has been since the very beginning of OpenSMCE.

Despite less clunky than UI1, it had its own flaws, but most importantly, it hasn't seen any adoption. And the fact that a lot of the files were needed and somewhat maintained to be kept in line with UI1 for UI2 to even work...? (UI2 wasn't even being tested!), which might have been causing too much extra work and also confusion as to why the seemingly superior UI version wasn't in use depite having 2 in its name, has been making UI2's position uncertain for a long time.

But the absolutely final nail to the coffin has been hammered in by SkilletUI, the new UI system which has seen a lot of development lately, and will at some point be included with OpenSMCE. It also comes with a pretty cool editor, so that's a very nice bonus!

So, UI2 is getting removed. This includes all documentation, schemas, Config Classes, files and also defintions in the Resource Manager and Documentation Generator.

There will be a few remnants of UI2 left here and there due to it sharing certain design paradigms with SkilletUI.

Changes in game config:
- The `useUI2` field has been removed.