! De-hardcode level start, multiple level sequences
Until now, when the level has been started by the UI script, the level state was hardcoded in a so-called "0-th step" which was not changeable. Now that step has been extracted to the level sequence file.
This change also grants the UI script ability to not reset the music when starting a level and to introduce per-level intro transitions.
Level sequences have been also moved out from `gameplay.json` into a new `level_sequences` folder.

Changes in `gameplay.json`:
- The `levelSequence` section has been removed.
  - Its contents should be moved to a level sequence file.

Changes in level data:
- Added a new required field:
  - `sequence` - The path to a Level Sequence which should be followed by this Level.

Added a new resource type: Level Sequences!
- These are the same level sequences that have been sitting in `gameplay.json` until now.
- Their home folder is `level_sequences/` but as usual you can store them anywhere you want, since it's the `$schema` field in the file itself which determines the actual file type.

Changes in Level Sequence data:
- Added a new entry type:
  - `"uiCallback"` - Executes a UI callback. Has two required fields:
    - `callback` - The UI callback which should be executed.
    - `waitUntilFinished` - If `true`, the sequence will wait until the `levelContinue` function is executed on the UI script side. Otherwise, the sequence will immediately move on to the next step.

Changes in UI Scripts:
- The `levelBegin` and `levelBeginLoad` functions have been removed.
- Added two new functions:
  - `levelRestartMusic` - Restarts the level music from the beginning.
  - `levelContinue` - Triggers the next level sequence step, but only when the level is currently waiting for that to happen.

In order to preserve the current behavior, you need to add the following entry to the level sequence, as the first entry (and move pointers to other entries appropriately):
```
{
    "type": "uiCallback",
    "callback": "levelStart",
    "waitUntilFinished": true,
    "muteMusic": true
}
```
and in the UI script change:
- `f.levelBegin()` to `f.levelRestartMusic()` and `f.levelContinue()`,
- `f.levelBeginLoad()` to `f.levelRestartMusic()`.