! Instant matches
A new experimental feature has been added: instant matches.
When enabled in the gameplay config file, the spheres will not be appended and instead match instantly if they can make a match.
If the match effect has a delay, the sphere will be attached, which is an experimental feature.

Changes in `gameplay.json`:
- Added a new optional field: `instantMatches` - If `true`, the spheres will match instantly when they can match. If the match effect has a delay, the sphere will be attached to the nearest sphere and will match after the specified period of time.