! `redirectedHitSphere` context
Commit #1000!!!

A new Expression Context called `redirectedHitSphere` has been added and will appear in Sphere Selectors referenced in the `hitBehavior` field of Spheres.
`redirectedHitSphere` is the same as `hitSphere` unless `hitSphere` is a scarab. In that case, `redirectedHitSphere` will contain the first sphere in front of the scarab.

Changes in Sphere data:
- The `color` field in the `recolorSpheres` hit behavior can be now an Expression.