! Homing Bugs
A new feature: Homing Bugs!
When it's active, any sphere you shoot will gravitate towards the biggest clump of that color.
When the target sphere is destroyed, the shot sphere will pick a new target, or be destroyed if no spheres are on the screen.

Changes in collectible effect data:
- Added a new collectible effect type: `"homingBugs"`. The shot spheres will gravitate towards the largest sphere group of matching color.
  - Contains one parameter:
    - `time` - The duration of the effect in seconds.