Experimental resource injection
This is an incomplete but working proof of concept for resource injection for Resource Managers.

The aim is so that the Documentation Generator can generate brand new files without changing existing ones (other than the files it has already generated) and automatically register resources which are instantly ready to use with Resource Managers.

More progress on this soon.