Code deprecation notices
A new option has been added to the Boot Screen: "Print deprecation notices".
When it is enabled, the ingame console will show code deprecation messages along with a pointer to the relevant line of code.
Setting up a deprecation notice is as easy as adding a `_Debug:deprecationNotice()` call in the relevant chunk of code. In a few places, this call has been added.
Duplicate messages are never shown.