! Simplify Expression Variable code
Now, all Expression Variables are stored as their string IDs, instead of a table structure whenever a dot `.` is in the variable name. This also means that now any amount of these is possible.
The debug screen at F4 will now display some content on the right side of the screen.

Changes in Expressions:
- The Level Timer Series entries are no longer accessible with `level.<timer_series_name>_length`, but instead by `level.<timer_series_name>.length`.