! Chained Spheres
Spheres can now have a chain level, which defaults to 0.
When the chain level is above 0, if the sphere is going to be destroyed with a Sphere Effect, it will have one chain level removed, instead of the sphere completely destroyed. This behavior is not configurable yet.
Chains can be rendered on the Spheres using a sprite with a condition which uses the new `[sphere.chainLevel]` Expression Variable. Make sure to provide a default value like so:
`"${[sphere.chainLevel|0] == 1}"`
because otherwise the game will crash when trying to render the sphere in other contexts, such as the shooter!
Also, fixed Expression Variables crashing when the default value provided was `false`.

Changes in Sphere data:
- In the `sprites` field entry, added a new optional field:
  - `conditions` - If specified, all conditions inside this list must pass. Otherwise that particular sprite will not be rendered.
    - All `[sphere.*]` Expression Variables which do not require a position are available there.
- Added two new optional fields:
  - `chainDestroyParticle` - A one-time particle packet which will spawn when a chain layer is stripped from the Sphere.
  - `chainDestroySound` - A sound event played when a chain layer is stripped from the Sphere.

Changes in Level data:
- In the `trainRules` field:
  - For the `"random"` and `"pattern"` types - directly inside the main field list, and for the `"waves"` type - inside the key definition, added a new optional field:
    - `chainChances` - A list of percentage probabilities for each of the chain levels for any sphere of that key to spawn with.
      - If one chance is rolled, the subsequent do not roll. For example, for the `[0.25, 0.25]` value:
        - 25% of spheres will spawn with one chain layer,
        - 18.75% (25% of the remaining 75%) of spheres will spawn with two chain layers,
        - The remaining 56.25% of spheres will spawn with no chain layers.

- Changes in Expressions:
  - Added a `[sphere.chainLevel]` Expression Variable to everywhere where the `sphere` or `hitSphere` Expression Variable Contexts are present.
  - Most of the `[sphere.*]` Expression Variables are now available when the Sphere is rendered (for the `conditions` field in the sprite entry data).