! More work on Resource Aliases
Sometimes, due to unpredictable loading order, a resource might refer to another resource using an alias which is not yet known to the Resource Manager. In such case, the Resource Manager assumes that the aliased resource will be loaded later and an entry is stored inside its unresolved alias table for later resolution.
Unresolved alias resolution is performed for a resource when it is retrieved using one of the `get*()` functions. If the once missing alias has since been loaded, the alias reference will be filled inside the resource config data and the unresolved alias entry will be removed. If the alias reference cannot be resolved, an error will be thrown instead.

The problem arises when there are multiple resources in the chain. Suppose resource A refers to resource B, and resource B refers to resource C by an alias. Resource B is loaded before resource C, so the alias is registered as an unresolved alias reference.
Now, resource A is obtained with a getter. The problem is that there is no direct information that resource A depends on a resource which has an unresolved alias. Thus, resource B can be accessed and resource C is inaccessible from resource B despite no error being thrown.

Because of this problem, Resource Aliases have been deemed an unimplementable feature will be removed next commit.

Warning! This commit contains some debug code.