! Layered sphere rendering
The Display class now can handle multiple layers and so its API has slightly changed:
- `Display:setCanvas()` now accepts an optional `layers` argument, which is a list of layer names. By default, one layer named `"MAIN"` is created.
- `Display:canvasStart()` and `Display:canvasStop()` have been removed and replaced by `Display:setLayer()` and `Display:draw()`.
- `Display:setLayer()` selects a layer to draw to. All drawing calls are targeting that specified layer until either `Display:setLayer()` or `Display:draw()` is called.
- `Display:draw()` compiles all layers into one image, draws it, and clears all layers to prepare for the next frame.
The current problem with this approach is that each layer is stored on a separate canvas, which means one drawcall for each layer. If there is a lot of layers, this gets very impractical and GPU taxing. This problem cannot be alleviated by atlasing, unlike regular sprites, due to potentially very large layer canvas sizes. Additionally, alpha blending is broken and also custom blend modes wouldn't work nicely with other layers. This is why this solution is temporary and will be rewritten soon.
Currently, sphere rendering and map rendering has been moved to this system. This greatly simplifies drawing code.
The game looks broken because of this change due to aforementioned blending issues. This will be fixed soon.
The layer list for Canvases is being taken from the `config/hud_layer_order.json` file. The engine will crash unless the following layers exist, preferably in this order:
- `GameBackground`
- `GamePieceHShadow`
- `GamePieceHidden`
- `GameBackgroundSprites`
- `GamePieceNShadow`
- `GamePieceNormal`
- `GameScores`

Other news:
- Added `---@type` documentation for some fields for Maps, Paths, Sphere Chains and Sphere Groups.
- Path Entity shadows now rotate alongside the actual sprite.
- Improved SphereEntity code formatting.
- Added a new `hidden` state to Sphere Entities.